extracted logic for Modal
