+++ src/hooks/feature/useAuth.ts
@@ -21,7 +21,6 @@ const useAuth = () => {
     })
       .then((res) => res.json())
       .then((data) => {
-        console.log(data);
         if (data.error) {
           setError(data.error);
         } else {
